% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.R
\name{available}
\alias{available}
\alias{cur_avail_languages}
\alias{cur_avail_regions}
\alias{cur_avail_categories}
\title{Available Variables}
\usage{
cur_avail_languages()

cur_avail_regions()

cur_avail_categories()
}
\description{
Fetch abailable variables to use in other calls.
}
\examples{
\dontrun{cur_avail_regions()}

}
