% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{currents_key}
\alias{currents_key}
\title{Setup}
\usage{
currents_key(key)
}
\arguments{
\item{key}{Your api key as from \url{https://currentsapi.services}.}
}
\description{
Setup your session by specifiying your API key to be used in subsequent calls.
}
\note{
You can specify \code{CURRENTS_API_KEY} in your \code{.Renviron}.
}
\examples{
\dontrun{
currents_key("xxXXxxXxXxXXXxxXx")
}

}
