% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{news}
\alias{news}
\alias{latest_news}
\alias{search_news}
\title{News}
\usage{
latest_news()

search_news(keyword = NULL, country = NULL, language = NULL,
  start_date = NULL, end_date = NULL, type = NULL)
}
\arguments{
\item{keyword}{Restict the search to articles that mention a specific keyword.}

\item{country}{An \code{ISO 2} country code, full list of valid codes is available at \url{https://currentsapi.services/documents}.}

\item{language}{An \code{ISO 2} language code, full list of valid codes is available at \url{https://currentsapi.services/documents}.}

\item{start_date, end_date}{Start and end dates, either as \code{Date} or \code{POSIX}.}

\item{type}{Type of content to return \code{1} for news, \code{2} for article and \code{3} for discussion content. If \code{NULL} all 3 types are choosen.}
}
\value{
A \link[tibble]{tibble}
}
\description{
Return latest news or specific articles.
}
\examples{
\dontrun{
currents_key("xxXXxxXxXxXXXxxXx")

news <- search_news("Trump")
}

}
