% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{news}
\alias{news}
\alias{cur_latest}
\alias{cur_search}
\title{News}
\usage{
cur_latest()

cur_search(
  keyword = NULL,
  country = NULL,
  language = NULL,
  start_date = NULL,
  end_date = NULL,
  category = NULL
)
}
\arguments{
\item{keyword}{Restict the search to articles that mention a specific keyword.}

\item{country}{An \code{ISO 2} country code, see \code{\link{cur_avail_regions}} for a full list.}

\item{language}{An \code{ISO 2} language code, see \code{\link{cur_avail_languages}} for a full lis.}

\item{start_date, end_date}{Start and end dates, either as \code{Date} or \code{POSIX}.}

\item{category}{Valid category, see \code{\link{cur_avail_categories}} for a full list.}
}
\value{
A \link[tibble]{tibble}
}
\description{
Return latest news or specific articles.
}
\examples{
\dontrun{
currents_key("xxXXxxXxXxXXXxxXx")

news <- search_news("Trump")
}

}
